fix(manage-variables): make some changes
