+++ src/services/dashboards/dashboard-customize/modules/dashboard-manage-variable-overlay/modules/DashboardManageVariableForm.vue
@@ -68,7 +68,7 @@ import {
 } from '@spaceone/design-system';
 
 export default defineComponent({
-    name: 'DashboardManageVariableOverlay',
+    name: 'DashboardManageVariableForm',
     components: {
         PButton,
         PFieldGroup,
